python/save_chess_games/requirements.txt
chess
google-cloud-firestore
pyfarmhash